Consolidated timestamps tests
